kanban/pyproject.toml
[project]
name = "kanban"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.6.1",
]
